update: 通報機能のリファクタリング
